Fix TMC current step-down underflow

+++ Marlin/src/feature/tmc_util.cpp
@@ -335,7 +335,7 @@
     template<typename TMC>
     void step_current_down(TMC &st) {
       if (st.isEnabled()) {
-        const uint16_t I_rms = st.getMilliamps() - (CURRENT_STEP_DOWN);
+        const int32_t I_rms = int32_t(st.getMilliamps()) - int32_t(CURRENT_STEP_DOWN);
         if (I_rms > 50) {
           st.rms_current(I_rms);
           #if ENABLED(REPORT_CURRENT_CHANGE)
